refactor: 주문 확인 페이지 변경
1. 주문 확인 페이지를 결제 확인 페이지로 변경 후 새로운 주문 확인 페이지 컴포넌트를 구현하는 형태로 변경

2. Navigation 컴포넌트의 Header 컴포넌트를 url에 따라 객체에서 반환하도록 변경
