--- src/pages/OrderConfirmPage/OrderConfrimPage.styled.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const OrderConfirmPageContainer = styled.section`
-  margin-top: 20px;
-  width: 100%;
-  display: flex;
-  justify-contents: center;
-  align-items: center;
-`;
