--- src/pages/OrderPage/OrderPageErrorFallback.tsx
@@ -1,15 +1,0 @@
-interface ErrorFallbackProps {
-  error: Error;
-}
-
-function OrderPageErrorFallback({ error }: ErrorFallbackProps) {
-  return (
-    <div>
-      <div>
-        <p>{error.message}</p>
-      </div>
-    </div>
-  );
-}
-
-export default OrderPageErrorFallback;
