+++ src/pages/index.ts
@@ -1,2 +1,3 @@
+export { default as PaymentConfirmPage } from './PaymentConfirmPage/OrderConfirmPage';
+export { default as ShoppingCartPage } from './ShoppingCartPage/ShoppingCartPage';
 export { default as OrderConfirmPage } from './OrderConfirmPage/OrderConfirmPage';
-export { default as OrderPage } from './OrderPage/OrderPage';
